pyproject.toml
[tool.poetry]
name = "boltlib"
version = "0.2.0"
description = "Bitcoin Lightning BoltCard (NTAG 424 DNA) Read/Write library"
authors = ["Titusz <tp@py7.de>"]
license = "LGPL-2.1-only"
readme = "README.md"
repository = "https://github.com/titusz/boltlib"
keywords=["bitcoin", "lightning", "boltcard", "nfc", "contactless"]
classifiers=[
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Lesser General Public License v2 (LGPLv2)",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Hardware :: Universal Serial Bus (USB) :: Smart Card",
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/titusz/boltlib/issues"


[tool.poetry.scripts]
boltcard = 'boltlib.cli:cli'

[tool.poetry.dependencies]
python = "^3.8"
loguru = "^0.6"
pyscard = "^2.0"
ndeflib = "^0.3"
click = "^8.1"
construct = "^2.10.68"
construct-typing = "^0.5.4"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
black = "*"
poethepoet = "*"
pyinstaller = "*"
setuptools = "*"
wheel = "*"

[tool.poe.tasks]
format = { cmd = "poetry run black .", help = "Format code with black" }
test = { cmd = "poetry run pytest", help = "Run tests" }
all = ["format", "test"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
